refactor(core/rust): move micropython's Error to micropython module
in preparation of micropython spin-off to a separate crate,
and error handling rework in the trezor_lib
