+++ core/embed/rust/src/micropython/runtime.rs
@@ -1,8 +1,6 @@
 use core::mem::MaybeUninit;
 
-use crate::error::Error;
-
-use super::ffi;
+use super::{error::Error, ffi};
 
 /// Raise a micropython exception via NLR jump.
 /// Jumps directly out of the context without running any destructors,
