+++ core/embed/rust/src/protobuf/error.rs
@@ -1,24 +1,17 @@
-use crate::{
-    error::{value_error, Error},
-    micropython::qstr::Qstr,
-};
+use crate::micropython::{error::Error, qstr::Qstr};
 
 pub const fn experimental_not_enabled() -> Error {
-    value_error!(c"Experimental features are disabled.")
+    Error::ValueError(c"Experimental features are disabled.")
 }
 
 pub const fn unknown_field_type() -> Error {
-    value_error!(c"Unknown field type.")
+    Error::ValueError(c"Unknown field type.")
 }
 
-pub fn missing_required_field(field: Qstr) -> Error {
-    Error::ValueErrorParam(c"Missing required field.", field.into())
+pub const fn missing_required_field(field: Qstr) -> Error {
+    Error::ValueErrorParam(c"Missing required field.", field.to_obj())
 }
 
-pub fn invalid_value(field: Qstr) -> Error {
-    Error::ValueErrorParam(c"Invalid value for field.", field.into())
-}
-
-pub const fn end_of_buffer() -> Error {
-    value_error!(c"End of buffer.")
+pub const fn invalid_value(field: Qstr) -> Error {
+    Error::ValueErrorParam(c"Invalid value for field.", field.to_obj())
 }
